[APM] Fix Rison issues (#25377)
* [APM] Fix Rison issues

* Fix CI

* Removed abstractions (and broke tests)

* Slight rework and small bugfix for query param handling

* Removed typo used for visibility testing oops

* Removes unused dep because linting doesnt work apparently
